refactor(apis): api 레이어 내 디렉터리 및 파일 네이밍 변경
1. ApiSchema를 ApiClient로 변경

2. MovieAPI를 MovieFetcher로 변경
